requirements.txt
openenv-core
pytest
requests
httpx
openai
pydantic
python-dotenv
torch